Add event to operator buttons and fn operation
Removed the clear function and add it to fn operation with new code that
make all variables undefined. And for operator string except for = will
assign the display text to first num and all the click operator to
let operator. The equal string condition remain to be solved on why it
does not work